#10 example app showing error: Multiple directives [mdSidenavngScrollbars, new/isolated scope] asking for {4} on: {5}
Description
This error occurs when multiple directives are applied to the same DOM element, and processing them would result in a collision or an unsupported configuration.

To resolve this issue remove one of the directives which is causing the collision.

Example scenarios of multiple incompatible directives applied to the same element include:

Multiple directives requesting isolated scope.
Multiple directives publishing a controller under the same name.
Multiple directives declared with the transclusion option.
Multiple directives attempting to define a template or templateURL.

https://docs.angularjs.org/error/$compile/multidir?p0=mdSidenav&p1=ngScrollbars&p2=new%2Fisolated%20scope&p3=%3Cmd-sidenav%20ng-scrollbars%3D%22%22%20class%3D%22md-sidenav-left%20md-whiteframe-z2%20md-closed%22%20md-component-id%3D%22left%22%20md-is-locked-open%3D%22$mdMedia(%27gt-md%27)%22%20tabindex%3D%22-1%22%3E